Move form actions classes to form_actions.py
